demos/demo-jupyter/pyproject.toml
[build-system]
requires = ["setuptools>=60.0"]
build-backend = "setuptools.build_meta"

[project]
name = "jupyter_n8s"
version = "0.0.1"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "requests==2.28.1",
    "jupyter_server==1.18.1",
    "jupyter==1.0.0",
    "websockets==10.3",
]